Moved HCContour and HCPath to Geometry. Fixed a memory leak in HCCondition. Adjusted documentationo
