+++ Source/Geometry/HCContourCurve.c
