+++ Source/Geometry/HCContourCurve.h
