+++ Source/Geometry/HCPath+SVG.c
