+++ Source/Geometry/HCPath+SVG.h
