+++ Source/HollowCore.h
@@ -13,26 +13,32 @@
 
 #include "Core/HCCore.h"
 #include "Core/HCObject.h"
+
 #include "Data/HCNumber.h"
 #include "Data/HCString.h"
 #include "Data/HCData.h"
+
 #include "Container/HCList.h"
 #include "Container/HCSet.h"
 #include "Container/HCMap.h"
+
+#include "JSON/HCJSON.h"
+
+#include "Thread/HCCondition.h"
+#include "Thread/HCLock.h"
+#include "Thread/HCThread.h"
+
 #include "Geometry/HCPoint.h"
 #include "Geometry/HCSize.h"
 #include "Geometry/HCRectangle.h"
+#include "Geometry/HCContour.h"
+#include "Geometry/HCPath.h"
+#include "Geometry/HCPath+Shapes.h"
+#include "Geometry/HCPath+SVG.h"
+
 #include "Graphic/HCColor.h"
 #include "Graphic/HCRaster.h"
 #include "Graphic/HCRaster+Draw.h"
 #include "Graphic/HCRaster+File.h"
-#include "Graphic/HCContour.h"
-#include "Graphic/HCPath.h"
-#include "Graphic/HCPath+Shapes.h"
-#include "Graphic/HCPath+SVG.h"
-#include "JSON/HCJSON.h"
-#include "Thread/HCCondition.h"
-#include "Thread/HCLock.h"
-#include "Thread/HCThread.h"
 
 #endif /* HollowCore_h */
